Split and refactor dbService.ts
