+++ forms-flow-rsbcservice/src/formsflow-rsbcservices.ts
@@ -1,4 +1,5 @@
 // export all services here
-import DBService from "./storage/dbService";
+import DBFetchService from "./storage/dbFetchServices";
+import DBInsertService from "./storage/dbInsertServices";
 
-export { DBService };
+export { DBFetchService,  DBInsertService};
